+++ src/main/java/de/maxhenkel/camera/Main.java
@@ -36,12 +36,16 @@ import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.fml.network.NetworkRegistry;
 import net.minecraftforge.fml.network.simple.SimpleChannel;
 import net.minecraftforge.registries.ObjectHolder;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.lwjgl.glfw.GLFW;
 
 @Mod(Main.MODID)
 public class Main {
 
     public static final String MODID = "camera";
+
+    public static final Logger LOGGER = LogManager.getLogger(MODID);
 
     public static SimpleChannel SIMPLE_CHANNEL;
     public static PacketManager PACKET_MANAGER;